requirements.txt
aws-cdk.core==1.40.0
aws_cdk.aws_ec2==1.40.0
aws_cdk.aws_eks==1.40.0
aws_cdk.aws_rds==1.40.0
aws_cdk.aws_efs==1.40.0
aws_cdk.aws_ecs_patterns==1.40.0
aws_cdk.aws_elasticache==1.40.0
aws_cdk.aws_efs==1.40.0
aws_cdk.aws_ssm==1.40.0
boto3==1.13.11
invoke==1.4.1
